Do garbage collection to all dirty segments
