+++ ftl/page/page-gc.c
@@ -298,16 +298,17 @@ ssize_t page_ftl_do_gc(struct page_ftl *pgftl)
  *
  * @param pgftl pointer of the page ftl
  * @param request pointer of the request
+ * @param gc_ratio ratio for garbage collecting targets
  *
  * @return number of erased segments
  */
 ssize_t page_ftl_gc_from_list(struct page_ftl *pgftl,
-			      struct device_request *request)
+			      struct device_request *request, double gc_ratio)
 {
 	ssize_t ret = 0;
 	size_t nr_segments, nr_gc_segments, idx;
 	nr_segments = device_get_nr_segments(pgftl->dev);
-	nr_gc_segments = (size_t)((double)nr_segments * PAGE_FTL_GC_RATIO);
+	nr_gc_segments = (size_t)((double)nr_segments * gc_ratio);
 	for (idx = 0; (ssize_t)idx >= 0 && idx < nr_gc_segments; idx++) {
 		ret = page_ftl_submit_request(pgftl, request);
 		if (ret) {
